+++ nli/src/main/scala/vlp/nli/Teller.scala
@@ -170,7 +170,7 @@ class Teller(sparkSession: SparkSession, config: ConfigTeller, pack: DataPack) {
     val xs = validationSummary.readScalar("Top1Accuracy").map(_._2)
     Scores(arch = config.modelType, encoder = config.encoderType, maxSequenceLength = config.maxSequenceLength, embeddingSize = config.embeddingSize, 
       encoderSize = config.encoderOutputSize, bidirectional = config.bidirectional, tokenized = config.tokenized, 
-      trainingScores = xs.takeRight(20), testScore = scores._2)
+      trainingScores = xs.takeRight(20), testScore = scores._2.toFloat)
   }
 
   /**
